Florida Man Antagonist Definition
Adds the Florida Man antagonist, the event to spawn them in and a define file with possible objectives.

Sometimes it ain't so bad
Sometimes it ain't so good
I takes it all in stride
'Cause my livin's in the woods
Ain't got much crime or violence
An' we never had no drugs
But the moon comes through the pines
In mason jars and gallon jugs.