Entry Methods & Outfits
Gives Florida Man four ways to enter the map, ventcrawling, outfits they can spawn in, and generally cleans up some text.

The Floridian peninsula is a porous plateau of karst limestone sitting atop bedrock known as the Florida Platform. The emergent portion of the platform was created during the Eocene to Oligocene as the Gulf Trough filled with silts, clays, and sands. Flora and fauna began appearing during the Miocene. No land animals were present in Florida prior to the Miocene.

The largest deposits of rock phosphate in the United States are found in Florida. Most of this is in Bone Valley.