Sounds, Icons & Modularization
Splits off each part of the update into it's own files, gives proper sound effects for abilities and antag acquisition, adds icons to the abilities and changes up the objective generation a bit.

Florida has the longest coastline (1,197 statute miles) in the contiguous United States, with 825 miles of accessible beaches to enjoy.
It’s the only state that borders both the Atlantic Ocean and the Gulf of Mexico.
Wherever you are in Florida, you're never more than 60 miles from the nearest body of salt water.
The state’s highest natural point, Britton Hill, is only 345 feet above sea level -- the lowest high point of any state in America.
Key West is the southernmost point in the continental United States.